fix compiler warning
git-svn-id: https://scm.gforge.inria.fr/authscm/jmcoutey/svn/pm2/trunk@18579 ee107e0b-c2fa-0310-842a-ec2723fb34f4

+++ nmad/schedulers/sched_opt/src/nm_so_strat_split_balance.c
@@ -387,7 +387,7 @@ strat_split_balance_packv(void *_status,
     }
 
     if(iov[i].iov_len <= NM_SO_MAX_SMALL) {
-      NM_SO_TRACE("PACK of a small iov entry - tag = %u, seq = %u, len = %u, offset = %u, is_last_chunk = %u\n", tag, seq, iov[i].iov_len, offset, last_chunk);
+      NM_SO_TRACE("PACK of a small iov entry - tag = %u, seq = %u, len = %ld, offset = %u, is_last_chunk = %u\n", tag, seq, (long)iov[i].iov_len, offset, last_chunk);
       /* Small packet */
       strat_split_balance_try_to_agregate_small(_status, p_gate, tag, seq, iov[i].iov_base, iov[i].iov_len, offset, last_chunk);
